Wifi and txRx chages
1. Casted to a larger signed integer type in txRx
2. Changed to a lambda function in WiFiManager

+++ src/_txRx.ino
@@ -311,9 +311,12 @@ int sendPacket(uint8_t *buf, uint8_t len)
 	// And Server is also capable of transmitting on more channels.
 	// So, not change to SF (arranged by server) or frequency, as long as it is RX1 or RX2.
 
+	// ORIGINAL:
 	uint32_t fff	= (uint32_t)(root["txpk"]["freq"].as<double>() * 1000000);
 	
-	if (abs(freqs[gwayConfig.ch].dwnFreq - fff) < 100000) {
+	// // Changed using the Gemini
+	if (abs(static_cast<int64_t>(freqs[gwayConfig.ch].dwnFreq - fff)) < 100000) {
+	// if (abs(freqs[gwayConfig.ch].dwnFreq - fff) < 100000) {
 		LoraDown.freq = (uint32_t) (freqs[gwayConfig.ch].dwnFreq) & 0xFFFFFFFF ;
 		if ((debug>=2) && (pdebug & P_TX)) {
 			mPrint("v sendPacket:: _STRICT_1CH="+String(_STRICT_1CH)+", fff="+String(fff)+", abs");
